NUTCH-1732: allow deleting non-parsable documents
Add CrawlDatum status "parse_failed"/"db_parse_failed", set status in
CrawlDbReducer.

Handle deletion in IndexerMapReduce.

Add unit tests.
