+++ src/java/org/apache/nutch/parse/ParseStatus.java
@@ -188,6 +188,10 @@ public class ParseStatus implements Writable {
   public boolean isSuccess() {
     return majorCode == SUCCESS;
   }
+  
+  public boolean isFailed() {
+    return majorCode == FAILED;
+  }
 
   /**
    * @return a String representation of the first argument,
